Moved the region query to its own file
