+++ server/schemas/queries/index.js
@@ -1,2 +1,3 @@
 export { getAllUsers, getSpecificUser, getMe } from "./user.js";
 export { getVarieties } from "./variety.js";
+export { getRegions } from "./region.js";
